Fix Remove for empty values on indexes

+++ accounts/pkg/indexer/index/disk/unique.go
@@ -93,6 +93,9 @@ func (idx Unique) Add(id, v string) (string, error) {
 
 // Remove a value v from an index.
 func (idx Unique) Remove(id string, v string) (err error) {
+	if v == "" {
+		return nil
+	}
 	searchPath := path.Join(idx.indexRootDir, v)
 	return os.Remove(searchPath)
 }
